dynamic method calling from config.xml (<options>) -- still need to add error handling

+++ vaultTable.py
@@ -1,32 +1,61 @@
-import os, json
+import os, json, sys
 from cryptography.fernet import Fernet
 import xml.etree.ElementTree as ET
 
 class vaultTable:
-    def updateVaultTable(self,keyName,pw):
-        tmp = self.getVaultTable()
-        tmp['passwords'][keyName] = pw
-        with open(self.vaultKeyPath,'rb') as mykey:
-            key = mykey.read()
-        f = Fernet(key)
-        #TO-DO: error handling -- make sure file is locked
-        with open(self.getVaultPath(),'w') as vault:
-            json.dump(tmp,vault)
-        self.encryptVaultTable()
+    def read(self,prompt):
+        if sys.version_info[0] == 2:
+            return raw_input(prompt)
+        else:
+            return input(prompt)
+            
+    def updateVaultTable(self):
+        if not os.path.exists(self.getVaultKeyPath()):
+            print('ERROR: No vault key found at ', self.getVaultKeyPath())
+        else:
+            siteAlias = self.read('site alias? (i.e. "reddit")\n')
+            tmp = self.getVaultTable()
+            pw = self.read('enter pw:\n')
+            tmp['passwords'][siteAlias] = pw
+            with open(self.getVaultKeyPath(),'rb') as mykey:
+                key = mykey.read()
+            f = Fernet(key)
+            #TO-DO: error handling -- make sure file is locked
+            with open(self.getVaultPath(),'w') as vault:
+                json.dump(tmp,vault)
+            self.encryptVaultTable()
 
+    def printVaultTable(self):
+        if os.path.exists(self.getVaultKeyPath()):
+            with open(self.getVaultKeyPath(),'rb') as mykey:
+                key = mykey.read()
+            f = Fernet(key)
+            with open(self.getVaultPath(),'rb') as encryptedVault:
+                encrypted = encryptedVault.read()
+            decrypted = f.decrypt(encrypted)
+            try:
+                print(json.dumps(json.loads(decrypted),sort_keys=True,indent=4))
+            except:
+                print(json.loads(decrypted))
+        else:
+            print('ERROR: No vault key found at ', self.getVaultKeyPath())
+            
     def getVaultTable(self):
-        with open(self.vaultKeyPath,'rb') as mykey:
-            key = mykey.read()
-        f = Fernet(key)
-        with open(self.getVaultPath(),'rb') as encryptedVault:
-            encrypted = encryptedVault.read()
-        decrypted = f.decrypt(encrypted)
-        return json.loads(decrypted)
+        if os.path.exists(self.getVaultKeyPath()):
+            with open(self.getVaultKeyPath(),'rb') as mykey:
+                key = mykey.read()
+            f = Fernet(key)
+            with open(self.getVaultPath(),'rb') as encryptedVault:
+                encrypted = encryptedVault.read()
+            decrypted = f.decrypt(encrypted)
+            return json.loads(decrypted)
+        else:
+            print('ERROR: No vault key found at ', self.getVaultKeyPath())
 
     def getVaultPath(self):
         if self.args.config:
-            if os.path.exists(args.config):
-                xml = ET.parse(args.config)
+            if os.path.exists(self.args.config):
+                xml = ET.parse(self.args.config)
             else:
                 xmlPath = self.getRelScriptPath() + '\config\\config.xml'
                 xml = ET.parse(xmlPath)
@@ -39,7 +68,7 @@ class vaultTable:
         return root[0].text
 
     def encryptVaultTable(self):
-        with open(self.vaultKeyPath,'rb') as mykey:
+        with open(self.getVaultKeyPath(),'rb') as mykey:
             key = mykey.read()
         f = Fernet(key)
         vaultPath = self.getVaultPath()
@@ -52,21 +81,37 @@ class vaultTable:
     def generateVaultKey(self):
         print('generating vault.key....')
         key = Fernet.generate_key()
-        with open(self.vaultKeyPath, 'wb') as mykey:
+        with open(self.getVaultKeyPath(), 'wb') as mykey:
             mykey.write(key)
 
     def createVaultTable(self):
-        tmp = { "passwords" : {"tmp":"none"} }
-        if not os.path.exists(self.getRelScriptPath() + '\.vault'):
-            os.mkdir(self.getRelScriptPath() + '\.vault')
-        with open(self.getRelScriptPath() + '\.vault\\vault.json','w+') as f:
-            json.dump(tmp,f)
-        f.close()
-        print('vault created at ' + self.getRelScriptPath() + '\.vault\n')
-        self.setVaultKeyPath(self.getArgs())
-        self.generateVaultKey()
-        self.encryptVaultTable()
-        
+        createVaultBool = True      
+        if os.path.exists(self.getRelScriptPath() + '\.vault\\vault.json'):
+            print('!!!! WARNING: a vault already exists at ' + self.getRelScriptPath() + '\\.vault\\vault.json')
+            inputMsg = 'Continue? This will overwrite vault existing at ' + self.getRelScriptPath() + '\\.vault\\vault.json [y/n]: '
+            option = self.read(inputMsg)
+            if option.lower() == 'y':
+                if os.path.exists(self.getVaultKeyPath()):
+                    os.remove(self.getVaultKeyPath())
+                os.remove(self.getRelScriptPath() + '\.vault\\vault.json')
+            elif option.lower() == 'n':
+                createVaultBool = False
+                print('Aborting vault creation....\n')
+            else:
+                createVaultBool = False
+                print('No option selected')      
+        if createVaultBool == True:
+            tmp = { "passwords" : {"tmp":"none"} }
+            if not os.path.exists(self.getRelScriptPath() + '\.vault'):
+                os.mkdir(self.getRelScriptPath() + '\.vault')
+            with open(self.getRelScriptPath() + '\.vault\\vault.json','w+') as f:
+                json.dump(tmp,f)
+            f.close()
+            print('vault created at ' + self.getRelScriptPath() + '\.vault\n')
+            self.setVaultKeyPath(self.getArgs())
+            self.generateVaultKey()
+            self.encryptVaultTable()
+    
     def setVaultKeyPath(self,args):
         if args.key == None:
             self.vaultKeyPath = self.getRelScriptPath() + '\.hide\\vault.key'
@@ -98,4 +143,3 @@ class vaultTable:
         self.setRelScriptPath()
         self.setArgs(args)
         self.setVaultKeyPath(self.getArgs())
-
